refactor note group to render Note
